feat: 갤러리 화면 다이나믹 컬러 적용
