+++ app/build.gradle
@@ -54,6 +54,7 @@ dependencies {
     implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.5.1'
     implementation 'androidx.navigation:navigation-fragment-ktx:2.5.3'
     implementation 'androidx.navigation:navigation-ui-ktx:2.5.3'
+    implementation 'androidx.palette:palette:1.0.0'
 
     //splash
     implementation 'androidx.core:core-splashscreen:1.0.0'
