+++ app/src/main/java/com/juniori/puzzle/adapter/PuzzleBindingAdapter.kt
@@ -1,78 +1,14 @@
 package com.juniori.puzzle.adapter
 
-import android.graphics.drawable.Drawable
-import android.widget.ImageView
 import android.widget.TextView
 import androidx.databinding.BindingAdapter
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
-import com.bumptech.glide.load.MultiTransformation
-import com.bumptech.glide.load.resource.bitmap.FitCenter
-import com.bumptech.glide.load.resource.bitmap.RoundedCorners
-import com.bumptech.glide.request.RequestOptions
-import com.bumptech.glide.request.target.CustomTarget
-import com.bumptech.glide.request.transition.Transition
 import com.juniori.puzzle.R
 import java.util.Calendar
 import java.util.Date
 
 private val calendar = Calendar.getInstance()
-
-private const val DRAWABLE_WIDTH = 120
-private const val DRAWABLE_HEIGHT = 120
-
-@BindingAdapter("setImage")
-fun setImage(view: ImageView, url: String?) {
-    if (url.isNullOrEmpty()) return
-
-    Glide.with(view.context)
-        .load(url)
-        .into(view)
-}
-
-@BindingAdapter("setDrawableLeft")
-fun setDrawableLeft(view: TextView, url: String?) {
-    if (url.isNullOrEmpty()) return
-
-    try {
-        val resourceId = url.toInt()
-        Glide.with(view.context)
-            .load(resourceId)
-            .into(object : CustomTarget<Drawable>(DRAWABLE_WIDTH, DRAWABLE_HEIGHT) {
-                override fun onResourceReady(
-                    resource: Drawable,
-                    transition: Transition<in Drawable>?
-                ) {
-                    resource.setBounds(0, 0, DRAWABLE_WIDTH, DRAWABLE_HEIGHT)
-                    view.setCompoundDrawables(resource, null, null, null)
-                }
-
-                override fun onLoadCleared(placeholder: Drawable?) {
-                    placeholder?.setBounds(0, 0, DRAWABLE_WIDTH, DRAWABLE_HEIGHT)
-                    view.setCompoundDrawablesWithIntrinsicBounds(placeholder, null, null, null)
-                }
-            })
-    } catch (e: Exception) {
-        Glide.with(view.context)
-            .load(url)
-            .into(object : CustomTarget<Drawable>(DRAWABLE_WIDTH, DRAWABLE_HEIGHT) {
-                override fun onResourceReady(
-                    resource: Drawable,
-                    transition: Transition<in Drawable>?
-                ) {
-                    resource.setBounds(0, 0, DRAWABLE_WIDTH, DRAWABLE_HEIGHT)
-                    view.setCompoundDrawables(resource, null, null, null)
-                }
-
-                override fun onLoadCleared(placeholder: Drawable?) {
-                    placeholder?.setBounds(0, 0, DRAWABLE_WIDTH, DRAWABLE_HEIGHT)
-                    view.setCompoundDrawablesWithIntrinsicBounds(placeholder, null, null, null)
-                }
-            })
-    }
-
-}
 
 @BindingAdapter("setAdapter")
 fun <T> setAdapter(view: RecyclerView, itemList: List<T>) {
@@ -109,15 +45,16 @@ fun setTime(view: TextView, date: Date) {
     }
 }
 
-@BindingAdapter("imageBytes")
-fun setImageBytes(view: ImageView, imageBytes: ByteArray) {
-    Glide.with(view.context)
-        .load(imageBytes)
-        .apply(
-            RequestOptions.bitmapTransform(
-                MultiTransformation(RoundedCorners(20), FitCenter())
-            )
+@BindingAdapter("setDisplayName")
+fun setDisplayName(view: TextView, name: String) {
+    view.text = if (name.isNotEmpty()) {
+        String.format(view.context.getString(R.string.display_name_format), name)
+    } else {
+        String.format(
+            view.context.getString(R.string.display_name_format),
+            view.context.getString(R.string.display_anonymous)
         )
-        .into(view)
+    }
 }
+
