+++ app/src/main/res/layout/fragment_othersgallery.xml
@@ -44,6 +44,7 @@
         android:id="@+id/recycle_other_gallery"
         android:layout_width="0dp"
         android:layout_height="0dp"
+        android:layout_marginHorizontal="@dimen/item_horizontal_margin"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
